renamed interceptors
aiohttp_quick -> aiohttp
requests_quick -> requests
urllib -> urllib_full
urllib_quick -> urllib
urllib3 -> urllib3_full
urllib3_quick -> urllib3
